matcher/requirements.txt
numpy==1.23.3
scikit-learn==1.1.2
nltk==3.7
spacy==2.3.5
kafka-python==2.0.2
Flask==2.2.2
cassandra-driver==3.25.0